Rewrite framework, add test bot
Follow bot is a bot that uses the prediction of how a car moves to try
and keep up with other vehicle.

Speed tester is useful to debug how Follow bot will approach a given
situation.

Improvements to the functions, like getting the middle of the throttle
values instead of the end, etc.
This means new data was gathered.

Data splitter is used after data is gathered to put it into 4 categories:
 - accelerating
 - breaking
 - boosting
 - coasting

 Partial rewrite to the visualizers so they use the new data layout